Apply PR review feedback: relocate and refactor tests for SecurityValidator
