+++ src-tauri/src/mcp/builtin/utils.rs
@@ -425,237 +425,3 @@ pub async fn load_session_tool_access(session_id: Option<&str>) -> SessionToolAc
         agent_id,
     }
 }
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    #[test]
-    fn test_path_validation() {
-        let temp_dir = std::env::temp_dir().join("mcp_test");
-        let validator = SecurityValidator::new_with_base_dir(temp_dir.clone());
-
-        // Valid paths
-        assert!(validator.validate_path("test.txt").is_ok());
-        assert!(validator.validate_path("./test.txt").is_ok());
-        assert!(validator.validate_path("subdir/test.txt").is_ok());
-        assert!(validator
-            .validate_path("attachments/docker_조사....md")
-            .is_ok());
-
-        // Absolute paths for read operations should be allowed
-        assert!(validator
-            .validate_path_for_read("/tmp/some-file.txt")
-            .is_ok());
-
-        // Invalid paths (directory traversal)
-        assert!(validator.validate_path("../test.txt").is_err());
-        assert!(validator.validate_path("../../etc/passwd").is_err());
-
-        // Invalid paths (absolute paths) - 새로 추가된 보안 검증
-        assert!(validator.validate_path("/etc/passwd").is_err());
-        assert!(validator.validate_path("/Users/test/file.txt").is_err());
-        assert!(validator.validate_path("/tmp/outside.txt").is_err());
-
-        // Invalid paths (Windows drive letters) - 추가된 검증
-        assert!(validator.validate_path("C:\\Windows\\System32").is_err());
-        assert!(validator.validate_path("D:\\secret.txt").is_err());
-
-        // Invalid paths (complex traversal attempts)
-        assert!(validator
-            .validate_path("./subdir/../../../etc/passwd")
-            .is_err());
-
-        // Windows 스타일 경로도 상대경로로 처리되지만, ".." 포함으로 차단됨
-        assert!(validator.validate_path("subdir\\..\\..\\Windows").is_err());
-    }
-
-    #[test]
-    fn test_normalize_path_separators() {
-        let windows_path = "C:\\Users\\user\\file.txt";
-        let normalized = SecurityValidator::normalize_path_separators(windows_path);
-        assert_eq!(normalized, "C:/Users/user/file.txt");
-
-        let mixed_path = "C:/Users\\user/file.txt";
-        let normalized = SecurityValidator::normalize_path_separators(mixed_path);
-        assert_eq!(normalized, "C:/Users/user/file.txt");
-
-        let unix_path = "/home/user/file.txt";
-        let normalized = SecurityValidator::normalize_path_separators(unix_path);
-        assert_eq!(normalized, "/home/user/file.txt");
-    }
-
-    #[test]
-    fn test_extract_filename() {
-        // Windows paths
-        let path = "C:\\Users\\user\\Downloads\\test.pdf";
-        let filename = SecurityValidator::extract_filename(path);
-        assert_eq!(filename, Some("test.pdf".to_string()));
-
-        // Unix paths
-        let path = "/home/user/downloads/test.pdf";
-        let filename = SecurityValidator::extract_filename(path);
-        assert_eq!(filename, Some("test.pdf".to_string()));
-
-        // Mixed separators
-        let path = "C:/Users/user\\Downloads\\test.pdf";
-        let filename = SecurityValidator::extract_filename(path);
-        assert_eq!(filename, Some("test.pdf".to_string()));
-
-        // Edge cases
-        let path = "test.pdf";
-        let filename = SecurityValidator::extract_filename(path);
-        assert_eq!(filename, Some("test.pdf".to_string()));
-
-        let path = "";
-        let filename = SecurityValidator::extract_filename(path);
-        assert_eq!(filename, Some("".to_string()));
-
-        let path = "C:\\Users\\";
-        let filename = SecurityValidator::extract_filename(path);
-        assert_eq!(filename, Some("".to_string()));
-    }
-
-    #[test]
-    #[cfg(unix)]
-    fn test_symlink_traversal() {
-        use std::os::unix::fs::symlink;
-        let temp_dir = std::env::temp_dir().join("mcp_symlink_test");
-        if temp_dir.exists() {
-            std::fs::remove_dir_all(&temp_dir).unwrap();
-        }
-        std::fs::create_dir_all(&temp_dir).unwrap();
-
-        let validator = SecurityValidator::new_with_base_dir(temp_dir.clone());
-
-        // Create a secret file outside
-        let secret_file = std::env::temp_dir().join("mcp_secret.txt");
-        std::fs::write(&secret_file, "secret").unwrap();
-
-        // Create a symlink inside base_dir pointing to secret file
-        let link_path = temp_dir.join("bad_link");
-        symlink(&secret_file, &link_path).unwrap();
-
-        // Try to access via symlink
-        // The validator should reject this because the CANONICAL path is outside base_dir
-        let result = validator.validate_path("bad_link");
-
-        assert!(result.is_err(), "Symlink traversal should be blocked");
-        if let Err(SecurityError::PathTraversal(msg)) = result {
-            assert!(msg.contains("resolves outside allowed directory"));
-        } else {
-            panic!("Expected PathTraversal error");
-        }
-
-        // Cleanup
-        if let Err(err) = std::fs::remove_dir_all(&temp_dir) {
-            eprintln!("Failed to remove test directory {:?}: {}", temp_dir, err);
-        }
-        if let Err(err) = std::fs::remove_file(&secret_file) {
-            eprintln!("Failed to remove test file {:?}: {}", secret_file, err);
-        }
-    }
-
-    #[test]
-    #[cfg(unix)]
-    fn test_symlink_base_dir_traversal() {
-        use std::os::unix::fs::symlink;
-
-        // temp root for this test
-        let temp_root = std::env::temp_dir().join("mcp_symlink_base_dir_test");
-        if temp_root.exists() {
-            std::fs::remove_dir_all(&temp_root).unwrap();
-        }
-        std::fs::create_dir_all(&temp_root).unwrap();
-
-        // real base directory and a symlink used as the validator's base_dir
-        let real_base = temp_root.join("real_base");
-        std::fs::create_dir_all(&real_base).unwrap();
-        let base_dir_link = temp_root.join("base_link");
-        symlink(&real_base, &base_dir_link).unwrap();
-
-        // Use the symlinked path as base_dir; new_with_base_dir should canonicalize it
-        let validator = SecurityValidator::new_with_base_dir(base_dir_link.clone());
-
-        // secret file outside the real base
-        let secret_file = std::env::temp_dir().join("mcp_secret_base_dir.txt");
-        std::fs::write(&secret_file, "secret").unwrap();
-
-        // symlink inside the real base that points to the secret file outside
-        let escaping_link = real_base.join("escaping_link");
-        symlink(&secret_file, &escaping_link).unwrap();
-
-        // From the validator's perspective this looks like a valid relative path,
-        // but its canonical resolution escapes the real base directory.
-        let result = validator.validate_path("escaping_link");
-
-        assert!(
-            result.is_err(),
-            "Symlink traversal via symlinked base_dir should be blocked"
-        );
-        if let Err(SecurityError::PathTraversal(msg)) = result {
-            assert!(msg.contains("resolves outside allowed directory"));
-        } else {
-            panic!("Expected PathTraversal error");
-        }
-
-        // Cleanup
-        if let Err(err) = std::fs::remove_dir_all(&temp_root) {
-            eprintln!("Failed to remove test directory {:?}: {}", temp_root, err);
-        }
-        if let Err(err) = std::fs::remove_file(&secret_file) {
-            eprintln!(
-                "Failed to remove secret test file {:?}: {}",
-                secret_file, err
-            );
-        }
-    }
-
-    #[test]
-    #[cfg(unix)]
-    fn test_symlink_traversal_nonexistent_file() {
-        use std::os::unix::fs::symlink;
-        let temp_dir = std::env::temp_dir().join("mcp_symlink_nonexist_test");
-        if temp_dir.exists() {
-            std::fs::remove_dir_all(&temp_dir).unwrap();
-        }
-        std::fs::create_dir_all(&temp_dir).unwrap();
-
-        let validator = SecurityValidator::new_with_base_dir(temp_dir.clone());
-
-        // Create an outside directory
-        let outside_dir = std::env::temp_dir().join("mcp_outside_dir");
-        if outside_dir.exists() {
-            std::fs::remove_dir_all(&outside_dir).unwrap();
-        }
-        std::fs::create_dir_all(&outside_dir).unwrap();
-
-        // Create a symlink inside base_dir pointing to outside directory
-        let link_path = temp_dir.join("bad_link_dir");
-        symlink(&outside_dir, &link_path).unwrap();
-
-        // Try to access a nonexistent file via the symlink
-        let result = validator.validate_path("bad_link_dir/nonexistent_file.txt");
-
-        assert!(
-            result.is_err(),
-            "Symlink traversal to nonexistent file should be blocked"
-        );
-        if let Err(SecurityError::PathTraversal(msg)) = result {
-            assert!(msg.contains("resolves outside allowed directory"));
-        } else {
-            panic!("Expected PathTraversal error");
-        }
-
-        // Cleanup
-        if let Err(err) = std::fs::remove_dir_all(&temp_dir) {
-            eprintln!("Failed to remove test directory {:?}: {}", temp_dir, err);
-        }
-        if let Err(err) = std::fs::remove_dir_all(&outside_dir) {
-            eprintln!(
-                "Failed to remove test outside directory {:?}: {}",
-                outside_dir, err
-            );
-        }
-    }
-}
